#156176350 Improve lumen implementation (#13)
* chore(routes): implemented Dingo API on routes

This commit adds Dingo API to assist for routing. It uses the
Dingo resource method to enable a cleaner routes file.

It also moves the controller to a V1 folder to ensure proper
API versioning. Namespace changes have also been done.

Methods in the Controllers have also been changed to match
those provided by the Dingo routing resource, thus methods
such as 'index', 'store', 'update', 'destroy' and 'show'
have been used to facilitate use of the dingo resource method.

[Delivers #156176350]

* chore(read-me): updates the read me

This commit updates the README for the repository to contain set up instructions
for the API. It also provides a listing of all the routes provided by the API.

[Finishes #156176350]

* chore(exceptions): adds the internal server error exception

This commit adds the internal server error exception to the
application to ensure that when it is thrown, it returns a
customised response to the user with the http error status
code 500 and the message which the developer will want to
return

[Finishes #156176350]